Restructure api and output files
