+++ src/imgui/tsconfig.build.json
@@ -1,7 +1,7 @@
 {
     "$schema": "https://json.schemastore.org/tsconfig.json",
 
-    "files": ["../../bindgen/mod.ts"],
+    "include": ["../../bindgen/ts/"],
     "compilerOptions": {
         "outDir": "../../build",
         "target": "ESNext",
